Changed List[int] to List[float]

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -5,7 +5,7 @@ from typing import List
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: int, max_delay: int) -> List[int]:
+async def wait_n(n: int, max_delay: int) -> List[float]:
     """ Returns list of wait times """
     tasks = [asyncio.create_task(wait_random(max_delay)) for _ in range(n)]
     return [await task for task in asyncio.as_completed(tasks)]
